Do bandwidth scheduler header upgrade the same way as for congestion control (#12307)
During review of bandwidth scheduler code, @wacban mentioned that he'd
prefer the header upgrade to be done the same way as it was done for
congestion control
[ref](https://github.com/near/nearcore/pull/12234#discussion_r1812553819),
but I wasn't convinced if it's really cleaner.

In this PR I modified the header upgrade to work the same way as it does
in congestion control. We can compare the two approaches and choose the
better one.

The current approach looks like this:
* Before protocol upgrade to `BandwidthScheduler` version all chunks use
`InnerV3`, which doesn't have bandwidth requests in it.
* After the protocol version upgrade all newly produced chunks should
have `InnerV4`. Application of the last chunk of the previous protocol
version will produce a `ChunkExtra` which doesn't have bandwidth
requests (they are set to `None`), so the bandwidth requests in
`InnerV4` of the first chunk are set to the default value. Bandwidth
requests in `InnerV4` are not an `Option`, so we can't set them to
`None`.
* After the first chunk all produced `ChunkExtras` will have
`bandwidth_requests` set to `Some`, and they'll be put inside `InnerV4`
* `validate_chunk_with_chunk_extra_and_receipts_root` needs to be aware
of what happens at the first block and allow situations where the
bandwidth requests in `ChunkExtra` are `None`, but they're
`Some(Default::default())` in the chunk header.

The congestion-control-like approach looks like this:
* Before protocol upgrade to `BandwidthScheduler` version all chunks use
`InnerV3`, which doesn't have bandwidth requests in it.
* The first chunk after the upgrade will still use `InnerV3` because the
`bandwidth_requests` in `ChunkExtra` are None.
* For future chunks the `bandwidth_requests` in `ChunkExtra` will be
`Some` and all chunks will use `InnerV4` with the requests.
* `validate_chunk_with_chunk_extra_and_receipts_root` can do a direct
comparison between the bandwidth requests in chunk extra and chunk
header.


In the current approach I like the exactness - all chunk headers
produced with the new protocol version have a new version of `Inner`. We
don't allow multiple header versions in one protocol version. The only
problem is that we need to have the special corner case check in
`validate_chunk_with_chunk_extra_and_receipts_root` . I think we also
need to make sure that we validate the header versions for endorsed
chunks, using `is_valid_for` or something like that.

The congestion-control-like approach doesn't have the weird corner case,
which is nice. It also doesn't require such strict validation of header
version - headers with wrong version will get rejected by chunk extra
validation because of None/Some difference. But it's much less exact. We
allow multiple inner versions for one protocol version, and I find that
much harder to reason about. I'm not sure what happens at genesis
chunks, it looks like we set the congestion infos to None, but that
means that genesis chunks would always have `InnerV2`, which would get
upgraded to `Inner<latest>` on the first chunk. weird. I changed them to
`Some(CongestionInfo::default())`, I think that makes things a bit
better, as now the chain starts with the current version of `Inner`.

Yet another approach would be to make bandwidth requests an `Option` in
`InnerV4`. They would be `None` on the first chunk and `Some` on the
next chunks. We could directly compare that with the requests in
`ChunkExtra`. But it's a bit sad that we'd have an `Option ` for
something that's supposed to always be there :/