feat(contract-distribution): Filter new deploys from accesses and add a new testloop test (#12298)
- Exclude the code-hashes from the newly deployed code from the message
sent from chunk-producer to chunk-validators. This is because code for
the newly deployed contracts are available in the state witness inside
the receipts (they are not excluded from the witness), and the witness
is already sent to the validators. Then the chunk validators will update
their contract cache while applying the deploy actions. Thus, the
validators do not need to request code for these.

- Add two testloop tests for contract distribution. A simple test
deploys and calls a contract. Another test deploys multiple contracts to
two different accounts/shards and calls them from different accounts. We
call the contracts before and after clearing the cache. We promote some
of the helper functions from the `congestion_control` test to the utils
to share with the new test.

- To support the test, we add a test-only method to the contract cache
to revoke the cache. This allows us to test that the chain can still
work if the cache is empty. Note that we only add the cleaning
functionality for testing, since the reasoning about cleaning the cache
while there are ongoing accesses is not trivial (for example see [this
comment](https://github.com/near/nearcore/pull/12300#discussion_r1814637453))

- We also change `env.rs` because the `warmup` step needs one more block
to ensure there is no missing chunk. The changes in the block heights in
`test_view_requests_to_archival_node` is related to this change.

---------

Co-authored-by: Andrea <Trisfald@users.noreply.github.com>