[Epoch Sync] Enable epoch sync by default in configs. (#12352)
The previous config defaulted to false, which is not good, because we
want all nodes to default it to true. This change renames the fields to
"disable_..." because (1) we need a field name change to apply new
defaults because some nodes may already have reinitialized the config;
(2) defaulting a field to false is more natural and less error-prone, so
using "disable" rather than "enable" will more naturally default to
enabling.

Also document the config fields.

Closes #12347