fix: remove recomputing memtrie node memory usages (#12362)
Issue: #12361

After #12359, we can take memory usages from `updated_nodes` instead of
recomputing them again.

The original plan was to remove reading child nodes completely, where it
is not necessary. However, we still need to read hash, and hash is still
a part of child' `MemTrieNodeId`... Moreover, it doesn't make sense to
store hashes of children in the node for memory savings.

I think we just need to live with it then. At least the usecase is very
clear now. If we ever want to make memtrie recording cleaner (make it a
side effect of `MemTrieNodePtr::view`?), we need to skip recording node
only if its hash is needed.