feat(resharding): flat storage resharding children catchup (#12312)
PR to add children catchup step for flat storages created as a result of
a parent shard split.

In previous iterations, the two children shards were populated in a
background task from the flat storage of the parent at height `last
block of old shard layout` (post-processing).

Since the task mentioned above takes a long time and the children are
active shards in the `first block of the new shard layout` their flat
storage accumulates a lot of deltas.

The catchup step applies delta in the background, then finalizes
creation of child flat storage, and triggers a possible memtrie rebuild.

Part of #12174