feat(contract-distribution): Add tests for scenarios that the same contract is deployed to different accounts (#12374)
The same contract may be deployed to different accounts. Without this
change, if we deploy a contract to account A and then B in different
chunks, we will attempt to distribute the contract after both chunks,
even if the contract code already distributed once.

We do not optimize this situation in this PR, but add more tests to
indicate the expected behavior. We add a new test to `apply.rs` which
demonstrates this case. Also we add two new simple testloop tests for
deploying and calling the same contract on two different accounts.