feat(contract-distribution): Avoid recording contract accesses when validating state witness (#12377)
When recording contract accesses we also check if the contract was in
the trie. This is needed when updating the shard (which leads to
generating the state witness). However, it leads to unnecessary trie
lookups and also corrupts the metrics.

To fix, this we first make `ApplyChunkReason` non-optional (this is not
necessary but better to always specify the reason instead of avoiding to
specify it in some cases). Then we skip the recording of contract
accesses if the reason indicates that we are applying the chunk for
validating the state witness.

We add a test to check that we do not record the access for validating
the witness.