feat: add validation for contracts code request (#12388)
We validate the requested contract hashes using stored state transition
data since chunk producer tracking the shard is expected to have it
available.
Also store already processed requests in the LRU cache to avoid sending
the same data multiple times to a malicious chunk validator.
`main_transition` is included as part of the message to avoid
recomputing it. Instead we could pass `prev_chunk_hash` and compute
shard_id the same way we when creating state witness, but that resulted
in much more complex implementation and a lot of code changes.
Note that we do not validate request's `main_transition` since there is
no way it can be abused. Even if chunk validator sends wrong data there
it still has to be a reasonably-sized set of contracts (since it
corresponds to a valid chunk application) and we only serve it once per
`ChunkProductionKey`.