fix(state-sync): delete struct ShardInfo (#12396)
`ShardInfo` stores a `ShardId` and a chunk header in the `StateSyncInfo`
stored on disk when we need to catch up. But the chunk header field is
not read anywhere, and wasn't even before
https://github.com/near/nearcore/pull/12102. Setting this chunk hash is
the first thing that fails when trying to enable state sync after a
resharding, because we try indexing into a block's chunks with the new
epoch's shard indices. So since it's totally unused, just remove it.

This makes a breaking change to the `StateDlInfos` column, but only
relative to another commit on master since the last release.