feat(resharding) - Testing non-contiguous ShardId part 1 (#12419)
Converted a few tests to the non-contiguous ShardId Layout. I'm pleased
to report that I found a bug and fixed it here too. Please see the
`build_receipts_hashes` method in chain.rs. It was relying on the
ShardIds being sorted and that broke some tests when using the shuffled
shard ids. Takeaway -> ShardId should not implement Ord / PartialOrd ->
I'll do that next. The remaining fixes are for test configuration.

This is to be continued - making the ShardIds really random breaks
another ~200 tests ;)