adjust state sync timeouts (#12425)
This PR improves configurability of state sync by separating the config
parameters for different kinds of timeouts:
- `state_sync_timeout` is used as it always has been when downloading
from external storage.
- `state_sync_p2p_timeout` is used when downloading from peers in the
network.
- `state_sync_retry_timeout` controls time between download attempts.

We also lower the number of retry attempts for p2p state sync from 5 to
3. Why? In testnet we see that requests are generally handled reliably
and if a request fails it is usually because the target node is
throttling incoming requests. In the case that peers are overwhelmed by
traffic we would rather fall back to cloud storage sooner than send more
requests (which will be throttled anyway).