state_witness: do not include new transactions with each state witness (#12270)
First and foremost this PR introduces an integration test that attempts
to push through a chunk that has a few invalid transactions. Previously
this would fail and make the entire chunk invalid.

A new protocol feature is introduced that adjusts the code in few
strategic places to remove the requirement for all transactions within a
chunk to be valid, instead delegating the responsibility of checking all
aspects of transaction validity to the moment when transactions get
converted to receipts.

Any transactions that are found to be invalid are simply ignored. At
this point all validators should agree equally as to which transactions
should be converted and which ones should be discarded.